+++ src/fdm_header.h
@@ -20,9 +20,7 @@ along with this program; if not, see <http://www.gnu.org/licenses/>.
 Author: Hans Bihs
 --------------------------------------------------------------------*/
 
-#include"vtu3D.h"
-#include"vtr3D.h"
-#include"vts3D.h"
+#include "printerCFD.h"
 
 #include"ioflow.h"
 #include"ioflow_void.h"
